moved board class representation to the right module
