+++ lib/src/features/board-centraline/models/board.dart
